+++ QuickShop-Bukkit/src/main/java/org/maxgamer/quickshop/converter/ApolloConfigurationConverter.java
@@ -1,0 +1,7 @@
+package org.maxgamer.quickshop.converter;
+
+public class ApolloConfigurationConverter {
+
+
+
+}
